Fix owner.id config type

+++ config/owner.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    'id' => env('OWNER_ID'),
+    'id' => (int)env('OWNER_ID'),
     'channel' => env('OWNER_CHANNEL'),
     'support' => env('OWNER_SUPPORT'),
 ];
